bit more re-org. broke out client and added useMemo
