--- components/examples/badge/src/IBadgeType.ts
@@ -1,12 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-import { IIconProps } from "@fluentui/react";
-
-export interface IBadgeType {
-    key: string;
-    text: string;
-    iconProps: IIconProps;
-}
