--- components/examples/badge/src/IHistory.ts
@@ -1,9 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export interface IHistory<T> {
-    value: T;
-    timestamp: Date;
-}
